Fix imports for all files - Rename 'AddParticipants' to 'AddDatasets' - Create 'index.tsx' for components/ folder
