+++ react/src/Admin/components/HidableTextButton.tsx
@@ -1,6 +1,5 @@
 import React, { useState } from "react";
-import { makeStyles } from "@material-ui/core/styles";
-import Button from "@material-ui/core/Button";
+import { makeStyles, Button } from "@material-ui/core";
 
 const useStyles = makeStyles(theme => ({
     root: {
