+++ react/src/Datasets/index.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from "react";
 import { makeStyles, Container, Grid } from "@material-ui/core";
-import DatasetTable from "./DatasetTable";
+import DatasetTable from "./components/DatasetTable";
 
 const useStyles = makeStyles(theme => ({
     appBarSpacer: theme.mixins.toolbar,
