+++ react/src/Settings/index.tsx
@@ -1,12 +1,14 @@
 import React, { useEffect, useState } from "react";
-import { makeStyles } from "@material-ui/core/styles";
-import Button from "@material-ui/core/Button";
-import Container from "@material-ui/core/Container";
-import Divider from "@material-ui/core/Divider";
-import Grid from "@material-ui/core/Grid";
-import Paper from "@material-ui/core/Paper";
-import TextField from "@material-ui/core/TextField";
-import Typography from "@material-ui/core/Typography";
+import {
+    makeStyles,
+    Button,
+    Container,
+    Divider,
+    Grid,
+    Paper,
+    TextField,
+    Typography,
+} from "@material-ui/core";
 
 const useStyles = makeStyles(theme => ({
     root: {
