+++ react/src/components/ConfirmModal.tsx
@@ -1,10 +1,13 @@
 import React from "react";
-import Button, { ButtonProps } from "@material-ui/core/Button";
-import Dialog from "@material-ui/core/Dialog";
-import DialogActions from "@material-ui/core/DialogActions";
-import DialogContent from "@material-ui/core/DialogContent";
-import DialogContentText from "@material-ui/core/DialogContentText";
-import DialogTitle from "@material-ui/core/DialogTitle";
+import {
+    Button,
+    ButtonProps,
+    Dialog,
+    DialogActions,
+    DialogContent,
+    DialogContentText,
+    DialogTitle,
+} from "@material-ui/core";
 
 export interface ConfirmModalProps {
     id: string;
