+++ react/src/components/ListItemRouterLink.tsx
@@ -1,8 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import ListItem from "@material-ui/core/ListItem";
-import ListItemIcon from "@material-ui/core/ListItemIcon";
-import ListItemText from "@material-ui/core/ListItemText";
+import { ListItem, ListItemIcon, ListItemText } from "@material-ui/core";
 
 export interface ListItemRouterLinkProps {
     primary: React.ReactNode;
