+++ react/src/components/Title.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Typography from "@material-ui/core/Typography";
+import { Typography } from "@material-ui/core";
 
 export default function Title(props: { children: React.ReactNode }) {
     return (
